Fix name clash of secret for application and route
We discovered that under specific circumstances a secret is created that
contains data for the route, albeit the secret is named just like the
application secret.

Turns out that there is obviously a name clash between the secret the
ACME controller creates for the route (Let's Encrypt certificate) and
the secret we create for the application in the pipeline.