Move instructions on testing to tests/README
Some instructions on running tests, specifically the ones on field
tests, were included in the CONTRIBUTING document, which is not a bad
place to include it, per se. However, other instructions on runnings the
test suite are included in tests/README, which makes related information
harder to find. It likely makes sense to move all related information
into a single place and point to there from all other places.