Migrate from py.path to pathlib in cookiecutter tests
The pytest-cookies Pytest plugin deprecated the `project` member of
the Cookiecutter execution result. The new `project_path` member is
a `pathlib.Path` object. The impact of this change is significant.

The `as_cwd()` context manager is not available in pathlib. We
compensate this loss by using the `cwd` argument of the `subprocess.run`
command. For convenience we use the cli-test-helpers implementation.